+++ Classes/Models/Character/Cell/ShootingCell/Projectile/ProjectileModel.h
@@ -10,14 +10,14 @@ class ProjectileModel {
 protected:
     LevelModel* level;
     ProjectileId id;
-    float cellX;
-    float cellY;
-    float directionVectorX;
-    float directionVectorY;
-    float targetX;
-    float targetY;
-    float speed;
-    float damage;
+    double cellX;
+    double cellY;
+    double directionVectorX;
+    double directionVectorY;
+    double targetX;
+    double targetY;
+    double speed;
+    double damage;
     CharacterModel* target;
     bool isDestroyed;
     virtual void effectOnHit(){};
@@ -25,8 +25,8 @@ public:
     static ProjectileModel* create(ProjectileId id, CharacterModel* source, CharacterModel* target);
     ProjectileModel(ProjectileId id, CharacterModel* source, CharacterModel* target);
     virtual void update();
-    virtual float getPositionCellX();
-    virtual float getPositionCellY();
+    virtual double getPositionCellX();
+    virtual double getPositionCellY();
     virtual CellPosition getPosition();
     virtual void hitTarget();
     virtual void __setLevel(LevelModel*);
