+++ Classes/Models/Character/CharacterModel.h
@@ -5,8 +5,8 @@
 class LevelModel;
 
 struct CellPosition {
-    float cellX, cellY;
-    CellPosition(float x, float y) {cellX = x; cellY = y;}
+    double cellX, cellY;
+    CellPosition(double x, double y) {cellX = x; cellY = y;}
 };
 
 class CharacterModel {
@@ -16,21 +16,21 @@ protected:
     int hp;
     CharacterStatus status;
     bool alive;
-    float cellX;
-    float cellY;
+    double cellX;
+    double cellY;
 public:
     CharacterModel(){};
     CharacterModel(CharacterType);
     virtual void takeDamage(int);
     virtual void setStatus(CharacterStatus);
     virtual bool isAlive();
-    virtual void setPosition(float, float);
+    virtual void setPosition(double, double);
     virtual CellPosition getPosition();
-    virtual float getPositionCellX();
-    virtual float getPositionCellY();
+    virtual double getPositionCellX();
+    virtual double getPositionCellY();
     virtual void update(){};
     virtual ~CharacterModel(){};
-    float getDistanceToOther(CharacterModel* other);
+    double getDistanceToOther(CharacterModel* other);
     virtual void __setLevel(LevelModel*);
 };
